Reorganize cuda/execution_policy.hpp underneath cuda/execution
