+++ agency/cuda.hpp
@@ -2,8 +2,7 @@
 
 #include <agency/detail/config.hpp>
 #include <agency/cuda/device.hpp>
-#include <agency/cuda/execution_policy.hpp>
-#include <agency/cuda/executor.hpp>
 #include <agency/cuda/future.hpp>
+#include <agency/cuda/execution.hpp>
 #include <agency/cuda/memory.hpp>
