+++ agency/cuda/execution/execution_policy.hpp
@@ -2,10 +2,10 @@
 
 #include <agency/detail/config.hpp>
 #include <agency/execution/execution_policy.hpp>
-#include <agency/cuda/executor/grid_executor.hpp>
-#include <agency/cuda/executor/parallel_executor.hpp>
-#include <agency/cuda/executor/concurrent_executor.hpp>
-#include <agency/cuda/executor/scoped_executor.hpp>
+#include <agency/cuda/execution/executor/grid_executor.hpp>
+#include <agency/cuda/execution/executor/parallel_executor.hpp>
+#include <agency/cuda/execution/executor/concurrent_executor.hpp>
+#include <agency/cuda/execution/executor/scoped_executor.hpp>
 #include <agency/detail/tuple.hpp>
 #include <type_traits>
