+++ agency/cuda/execution/executor/block_executor.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <agency/cuda/executor/grid_executor.hpp>
+#include <agency/cuda/execution/executor/grid_executor.hpp>
 #include <agency/detail/tuple.hpp>
 #include <agency/detail/invoke.hpp>
 #include <agency/detail/type_traits.hpp>
