+++ agency/cuda/execution/executor/multidevice_executor.hpp
@@ -1,11 +1,12 @@
 #pragma once
 
 #include <agency/detail/config.hpp>
-#include <agency/cuda/executor/grid_executor.hpp>
-#include <agency/cuda/executor/parallel_executor.hpp>
 #include <agency/execution/executor/executor_array.hpp>
 #include <agency/execution/executor/flattened_executor.hpp>
+#include <agency/cuda/execution/executor/grid_executor.hpp>
+#include <agency/cuda/execution/executor/parallel_executor.hpp>
 #include <numeric>
+#include <algorithm>
 
 namespace agency
 {
