--- agency/cuda/executor.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <agency/detail/config.hpp>
-#include <agency/cuda/executor/block_executor.hpp>
-#include <agency/cuda/executor/concurrent_executor.hpp>
-#include <agency/cuda/executor/grid_executor.hpp>
-#include <agency/cuda/executor/multidevice_executor.hpp>
-#include <agency/cuda/executor/parallel_executor.hpp>
-#include <agency/cuda/executor/scoped_executor.hpp>
-
